Time out psutil calls that take too long
Sometimes certain processes can have up to 1 million open files. This causes
dcrpm to hang. Instead, we should time out those calls. While it's not great,
chances are that even if said process is holding open the rpm lockfiles for
some reason, it's not worth it to wait to `stat()` all those files. Just
ignore that file and move on to the rest of them.