Granular index rebuilds and end to end testing
Summary:
First pass on attempting to selectively rebuild rpmdb indexes (on the premise that some rpm commands blow up on very specific indexes getting out of sync, examples in T28608827).
Complimentary test harness with real rpmdbs taken from prod and stashed in everstore.

The controller you requested could not be found., until we move this out of fbcode and have github be the source of truth

Reviewed By: jaymzh

Differential Revision: D7783857

fbshipit-source-id: 1aa73e0305a0b0a6fc4763f99fc758283a171a76