Run yum clean
Summary:
Apparently yum clean can fail due to rpmdb index inconsistencies. This might be already covered by other checks (index, and table checks), but let's have it here for completeness sake.

We already run yum clean in chef. BUT - we run with `--disablerepo=os,updates,fasttrack,extras,epel,debuginfo` - do we want to carry that over? It'd need to be a config option, because open source.

Also - yum is a fucking clown, and you cannot instruct it to use specific rpmdb via cmd line options. It does respect `~/.rpmmacros` and `%_dbpath`, but that's terrible pattern to use here - we're not gonna mess with users env. This means `--run-yum-clean` will be incompatible with `--dbpath` in dcrpm.

Reviewed By: jaymzh

Differential Revision: D8318949

fbshipit-source-id: fc0faba5320d10beb190f74748b61a95e0c8492d