Trigger database rebuild if recovery was unsuccessful
Summary:
Fixed error handling where the specific exception handling was hidden by generic exception.
Also added a db sanity check after recovery attempt to trigger db rebuild if necessary
This fixes an issue which db_recover was unable to do so.
The controller you requested could not be found., until we move this out of fbcode and have github be the source of truth

Reviewed By: michel-slm

Differential Revision: D8734780

fbshipit-source-id: ea7832d7d06607a7a58df0bc56f1e40d70301d9a