Kill any spinning rpm query processes
Summary:
We've seen these rpm query processes linger and burn CPU on
the web tier.  They end up being a non-trivial gCPU cost.  This diff
adds logic to dcrpm to find any rpm process that looks like a query
and, if it has been running more than an hour, kill it.

This is a dcpm implementation of D8703575.

Reviewed By: patcox

Differential Revision: D9237405

fbshipit-source-id: 09b139f6ca164729ecf849cc2e08765570bae9b7