Fix unit tests
Summary:
Unit tests use special logger TestHandler, which records all the actions to a list, and then, the test compares the list with a list of expected actions.
The problem is, the Handler, thus the list is shared between tests, so, next test inherits the list of action taken by previous tests.

So, we should clean the list, before starting a new test.

Reviewed By: joshuamiller01

Differential Revision: D9383153

fbshipit-source-id: 08bf7df2f809f14e84b7e5eaf8233efb43d65755