Only perform package count check on Linux
`dcrpm` considers the RPM database broken if the number of installed packages is below the threshold (currently 50).

This makes sense on Linux distributions that natively use RPM for package management (CentOS, Fedora) but not on
other platforms, for example macOS, where the total package count might legitimately be below this threshold.

Bypass the package count check if `dcrpm` is not running on Linux.