Switch the recommended test runner to pytest
`python setup.py test` triggers a weird behavior with setuptools >= 60,
which is not encountered when using `pytest`.

Using the latter also allows downstream distributions to easily exclude
certain tests as necessary, and makes debugging easier.

This requires a fix as pytest's logging code would fail to print the PID
of the mocked process, as that property was unset.

Resolves #47.

Signed-off-by: Michel Alexandre Salim <salimma@fedoraproject.org>